requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
jupyter
telethon
python-dotenv
tabulate
requests
pyyaml
vaderSentiment
emoji>=2.8.0
transformers>=4.44.0
# Torch pinning: use 2.3.1 on Python <3.13 (known-good on macOS), and a compatible newer torch on Python >=3.13
torch==2.3.1; python_version < "3.13"
torch>=2.7.1; python_version >= "3.13"
datasets>=2.20.0
accelerate>=0.26.0